requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.26.35
botocore==1.29.35
cachetools==5.2.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.2
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.2.0
Django==3.2
django-allauth==0.52.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-filter==22.1
django-formtools==2.4
django-oauth-toolkit==2.2.0
django-on-heroku==1.1.2
django-otp==1.1.4
django-phonenumber-field==6.4.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-storages==1.13.1
django-templated-mail==1.1.1
django-timezone-field==5.0
django-two-factor-auth==1.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-social-oauth2==1.2.1
drf-yasg==1.21.4
ecdsa==0.18.0
ez-setup==0.9
fancycompleter==0.9.1
google==3.0.0
google-api-core==2.11.0
google-auth==2.16.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
image-slicer==0.1.1
incremental==22.10.0
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.4.2
kombu==5.2.4
MarkupSafe==2.1.1
mpmath==1.2.1
msgpack==1.0.4
mysqlclient==2.1.1
numpy==1.23.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
patchify==0.2.3
pdbpp==0.10.3
phonenumbers==8.13.4
Pillow==9.3.0
prompt-toolkit==3.0.31
protobuf==4.21.12
psycopg2-binary==2.9.5
pushover @ git+https://github.com/Wyattjoh/pushover@5852545c5b9cf2717e1eafc4c8b134a08b0994da
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pypi==2.1
pyrepl==0.9.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-pushover==0.4
python3-openid==3.2.0
pytz==2022.4
qrcode==7.3.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
service-identity==21.1.0
simplejson==3.18.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
split-image==2.0.1
sqlparse==0.4.3
stripe==5.0.0
sympy==1.11.1
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.3.0
wmctrl==0.4
wrapt==1.14.1
zope.interface==5.5.1
